refactor: rename exportedtf to genhcl
